Refactor GamesProvider & move it to 'games'.
GamesProvider is much more pure and doesn't do any writing by itself,
has to be controlled by somebody else.

Move the control concern to FullProviderImpl.
New issue: Challonge Clanwars to be safe.
Added a safety net for ChallongeService as we don't currently have a barrier.
This is not enough to prevent false triggering of Challonges.